ref(onboarding): Split python gcpfunctions onboarding docs (#103049)
Contributes to
https://linear.app/getsentry/issue/TET-864/introduce-folders-for-onboarding-platforms
