+++ static/app/gettingStartedDocs/python/gcpfunctions/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './gcpfunctions';
+import docs from '.';
 
 describe('gcpfunctions onboarding docs', () => {
   it('renders doc correctly', () => {
